Fixed rest of renaming and namespace structure
