+++ ruby_event_store_rom_sql/README.md
@@ -1,5 +1,7 @@
-# RubyEventStoreRomSql
+# RubyEventStore::ROM
 
  A ROM based implementation of events repository for [Ruby Event Store](https://github.com/RailsEventStore/rails_event_store).
 
-This is an alternative repository used in `rails_event_store` gem.
+This is an alternative ActiveRecord `EventRepository` used in `rails_event_store` gem.
+
+The Ruby Event Store adapter for [rom-sql](https://github.com/rom-rb/rom-sql) is `RubyEventStore::ROM::EventRepository`.
