+++ ruby_event_store_rom_sql/lib/ruby_event_store/expected_version.rb
@@ -1,4 +1,4 @@
-module RubyEventStoreRomSql
+module RubyEventStore
   class ExpectedVersion
     POSITION_DEFAULT = -1.freeze
     NOT_RESOLVED = Object.new.freeze
