+++ ruby_event_store_rom_sql/lib/ruby_event_store/index_violation_detector.rb
@@ -1,4 +1,4 @@
-module RubyEventStoreRomSql
+module RubyEventStore
   class IndexViolationDetector
 
     MYSQL_PKEY_ERROR    = "for key 'PRIMARY'"
