+++ ruby_event_store_rom_sql/lib/ruby_event_store/rom/mappers/serialized_record.rb
@@ -1,6 +1,6 @@
 require 'rom/transformer'
 
-module RubyEventStoreRomSql
+module RubyEventStore
   module ROM
     module Mappers
       class SerializedRecord < ::ROM::Transformer
