+++ ruby_event_store_rom_sql/lib/ruby_event_store/rom/relations/event_streams.rb
@@ -1,6 +1,6 @@
 require 'securerandom'
 
-module RubyEventStoreRomSql
+module RubyEventStore
   module ROM
     module Relations
       class EventStreams < ::ROM::Relation[:sql]
