+++ ruby_event_store_rom_sql/lib/ruby_event_store/rom/relations/events.rb
@@ -1,6 +1,6 @@
 require 'yaml'
 
-module RubyEventStoreRomSql
+module RubyEventStore
   module ROM
     module Relations
       class Events < ::ROM::Relation[:sql]
