+++ ruby_event_store_rom_sql/lib/ruby_event_store/rom/repositories/events.rb
@@ -1,4 +1,4 @@
-module RubyEventStoreRomSql
+module RubyEventStore
   module ROM
     module Repositories
       class Events < ::ROM::Repository[:events]
