+++ ruby_event_store_rom_sql/lib/ruby_event_store/rom/version.rb
@@ -1,3 +1,5 @@
-module RubyEventStoreRomSql
-  VERSION = "0.1.0"
+module RubyEventStore
+  module ROM
+    VERSION = "0.1.0"
+  end
 end
